+++ sdk/fabric/Azure.ResourceManager.Fabric/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release History
 
+## 1.1.0-beta.4 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 1.1.0-beta.3 (2026-08-20)
+
+### Bugs Fixed
+
+- Fixed an ambiguous `ArmFabricModelFactory.FabricCapacityPatch` overload for callers passing null administration members, by renaming one of the overloads `ArmFabricModelFactory.FabricCapacityPatchWithProperties`.
+
 ## 1.1.0-beta.2 (2026-08-18)
 
 ### Features Added